Bugfix for mysterious Subject heading indicator
Catches errors where ind2 in the 6XX fields is not 1-7 and is not
missing. According to the MARC standard these should not exist, but
values including 8 have been encountered. This update merely
catches the errors and logs them, and investigation is needed into
the significance and meaning of these indicators