SFR-321 Implement better URI parsing of holdings (856) data
This implements a more robust parsing strategy for data in the 856
fields of OCLC catalog records. Previously some checking was done
to resolve links to `Item`s if possible. This implements more
checking for such items through a regex.

Additionally this also checks the URIs provided to see if an
identifier (such as OCLC or Gutenberg #) can be extracted. If so,
the URI is transformed into an identifier for current instance, as
that should provide better matching to existing records.

Additionally this refactored the OCLC parser somewhat to enable
better unit testing, which was also extended in this commit.