Add Sharding to Output Method
This adds the ability to have multiple shards that are read by the
database update function. Similar to the Classify service, this uses
the first oclc identifier associated with each record to generate the
partitionKey, which provides some grouping (though this should be
irelevant).

This makes the ingest pipeline more scalable and should generally
allow for faster processing of records.