Handle errored requests to OCLC Lookup
At present the OCLC lookup service can respond with a range of unclear/opaque errors, including 104 Connection Reset and others. This sets a short timeout for the initial request (2 seconds, when successful processing usually takes at most 750ms). If this times out or recieves a `ConnectionError` this will retry once at a longer timeout before failing. This should a) improve the success rate for OCLC lookup queries and b) lower the error rate for the lambda function