SFR-714 Update ebook fetching from 856 field
Previously this field attempted to find and include any reference to an ebook link that was present in the `856` field. It did so by both applying a regex to the `u` subfield with the URI,  and the `k` and `q` subfields, which contain descriptions of the links. This instead applies the concept of a whitelist to only check for the presence of links from specific providers, namely: HathiTrust, InternetArchive and Project Gutenberg.

This checks a dict of regex values to evaluate the URI and if a match is found, store it. This has the side effect of simplifying the execution of this function and ensuring that external editions are only fetched in the expected situations