SFR-703 Improve stream message ordering
This update ensures that instance records are placed in the same shard as their parent work as identified by the UUID of that work. This is intended to avoid deadlocks by ensuring the updates for a single work are executed sequentially. This should also have the side effect of preventing race conditions (which are already handled, but which should be more effecient if they are handled in this way). This expects to recieve the work UUID as a field in the incoming queue message and so must be updated after the OCLC classify function.