SFR-703 Make execution internal to classify process
Previously the ResearchNow ETL pipeline loaded work records from the Classify endpoint and then subsequently fetched the instance records from the lookup API asynchronously. This was done to ensure speedy execution while the Classify service was reading from a kinesis stream and the concurrency was limited, meaning that fast execution at that stage was vital. The drawback to this was that it requires a doubling of the update operations in the database in order to first store the "stump" instance records from the Classify function and then the fuller instances from the lookup function.

The Classify function has since been converted to read from a SQS queue, which allows a much greater degree of concurrency. This means that these invocations can be longer running with a much lower performance hit (if there is one at all). Because of this it should be possible to do the fetching of instance records synchronously, and returning these results as a unified object for persistence in the database. This reduces the number of lookups and write operations in the db, potentially resulting ina large performance gain. It also eliminates the possibility of race conditions occuring with the instance records.

To do so this function will not read from a queue but be invoked via an API call from the Classify function (in a multithreaded invocation), this makes these changes. It makes no changes to the internal logic, but represents a major change in how the function is invoked and returns data. This returns an object that shares its formatting with the VIAF function, so consuming the response should be very straightforward and familiar.