SFR-703 Add local agent authority fetch
To further isolate the logic of creating the records form the database layer this adds the method to fetch authority numbers from agents to this function. This should reduce the number of times this is invoked in the database layer. With greater concurrency at this point it makes more sense to do here.