SFR-693 Exclude restricted IA books
The Internet Archive's collection includes works that are not borrowable without an account (generally these are post-1923 editions of works with editions in the public domain). Since we cannot provide these to users they should be excluded from SFR results.

This can be done with a simple check against the metadata associated with each IA work, excluding those marked as restricted. No further changes as necessary as this prevents them being loaded from the source with regards to the rest of the SFR data pipeline.