SFR-802 Add Parsing of HathiTrust catalog links
The OCLC lookup service frequently returns HathiTrust "catalog" links in the 856 field of its MARC records. These URLs point to pages that catalog all of the items for a specific instance, which is good, but not particularly useful for our target audience.

This takes these links and transforms them into lists of links that can take users directly to readable copies of these books (or provide access to a downloadable PDF.)

To do so the methods govering parsing the 856 field have been moved into a separate class as this is by far the most complex aspect of parsing the MARC fields. It raises an error that is handled in the events that it cannot properly process an 856 field and, as before, it adds a link directly to an instance record if it cannot be determined to lead to a readable version of the work being enhanced.